requirements.txt
opencv-python==3.4.2.17
opencv-contrib-python==3.4.2.17
flask==1.1.1
flask_cors==3.0.8
numpy==1.17.2
Pillow==6.1.0
torch==1.1.0
torchvision==0.2.2